+++ src/app/apply-wanted-senior/(components)/(steps)/Info.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import styled from 'styled-components';
 import { WISH_SENIOR_MENTOR_MSG } from '../../constant';
 import NextBtn from '@/components/Button/NextBtn';
-import { WishSeniorTitle, WishSeniorSubTitle, NextBtnBox } from '../../page';
+import { WishSeniorTitle, WishSeniorSubTitle, NextBtnBox } from '../(template)';
 
 export function WishSeniorInfo({ onClick }: { onClick: () => void }) {
   return (
