+++ src/app/apply-wanted-senior/(components)/(steps)/Professor.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { WISH_SENIOR_MENTOR_MSG } from '../../constant';
-import { NextBtnBox, WishSeniorSubTitle, WishSeniorTitle } from '../../page';
+import { NextBtnBox, WishSeniorSubTitle, WishSeniorTitle } from '../(template)';
 import TextForm from '@/components/SingleForm/TextForm';
 import { useForm } from 'react-hook-form';
 import NextBtn from '@/components/Button/NextBtn';
