+++ src/app/apply-wanted-senior/(components)/(steps)/Submit.tsx
@@ -1,5 +1,5 @@
 import NextBtn from '@/components/Button/NextBtn';
-import { WishSeniorSubTitle, WishSeniorTitle } from '../../page';
+import { WishSeniorSubTitle, WishSeniorTitle } from '../(template)';
 import { BtnStyleNonM } from '@/components/Button/NextBtn/NextBtn.styled';
 import Image from 'next/image';
