+++ src/app/apply-wanted-senior/(components)/(template)/index.tsx
@@ -1,0 +1,19 @@
+import { styled } from 'styled-components';
+
+export const WishSeniorTitle = styled.h3`
+  font-size: 17px;
+  font-weight: 650;
+  color: #000000;
+`;
+
+export const WishSeniorSubTitle = styled.h4`
+  font-size: 14px;
+  font-weight: 550;
+  color: #6d747e;
+`;
+
+export const NextBtnBox = styled.div`
+  position: absolute;
+  top: 35rem;
+  width: 330px;
+`;
