+++ src/app/apply-wanted-senior/page.tsx
@@ -139,21 +139,3 @@ export default function ApplyWantedSeniorPage() {
     </main>
   );
 }
-
-export const WishSeniorTitle = styled.h3`
-  font-size: 17px;
-  font-weight: 650;
-  color: #000000;
-`;
-
-export const WishSeniorSubTitle = styled.h4`
-  font-size: 14px;
-  font-weight: 550;
-  color: #6d747e;
-`;
-
-export const NextBtnBox = styled.div`
-  position: absolute;
-  top: 35rem;
-  width: 330px;
-`;
